+++ R/assertions.R
@@ -1,0 +1,3 @@
+assert_fselector = function(fselector) {
+  assert_r6(fselector, "FSelector")
+}